First working version. Only d/i specifiers without any flags

+++ libft/ft_abs.c
@@ -6,13 +6,13 @@
 /*   By: aimelda <aimelda@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/13 00:35:42 by aimelda           #+#    #+#             */
-/*   Updated: 2019/09/13 00:39:09 by aimelda          ###   ########.fr       */
+/*   Updated: 2020/02/09 12:27:03 by aimelda          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_abs(int n)
+size_t	ft_abs(long long n)
 {
 	if (n < 0)
 		return (n * (-1));
